Switch back to a `splitbutton` instead of a `menubutton`
This patch changes the behavior of the insert menu dropdown, whereas
previously we had a separate add media button and then the insert
button, which always opened the drop-down. Now, we have a single button
that will add media on a direct click and only open the drop-down when
the chevron is clicked.